# app/swagger_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Doc Manager API",
    "description": "Swagger/OpenAPI 문서의 버전 관리 및 변경사항 추적 시스템 API",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "API 서버"
    }
  ],
...
  "paths": {
...
    "/urls/{urlId}/versions/latest": {
      "get": {
        "tags": [
          "Versions"
        ],
        "summary": "최신 버전 조회",
        "description": "최신 버전을 조회합니다. `swaggerJson`이 포함되어 있어 Swagger UI에 바로 표시할 수 있습니다. 버전이 없는 경우 404 에러가 반환됩니다.",
        "operationId": "getLatestVersion",
        "parameters": [
          {
            "name": "urlId",
            "in": "path",
            "required": true,
            "description": "API URL의 MongoDB ObjectId",
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$",
              "example": "507f1f77bcf86cd799439011"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "_id": "507f1f77bcf86cd799439012",
                    "urlId": "507f1f77bcf86cd799439011",
                    "versionId": "v3",
                    "versionNumber": 3,
                    "timestamp": "2024-11-27T10:00:00.000Z",
                    "swaggerJson": {
                      "openapi": "3.0.0",
                      "info": {
                        "title": "User Service API",
                        "version": "1.0.0"
                      },
                      "paths": {}
                    },
                    "endpointCount": 15,
                    "parameterCount": 25,
                    "summary": "5개 추가, 2개 삭제, 3개 수정"
                  }
                }
              }
            }
          },
          "404": {
            "description": "버전을 찾을 수 없음 (버전이 아직 생성되지 않음)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "NOT_FOUND",
                    "message": "버전을 찾을 수 없습니다"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/versions/latest": {
      "get": {
        "tags": [
          "Versions"
        ],
        "summary": "모든 URL의 최신 버전 목록",
        "description": "모든 URL의 최신 버전 목록을 조회합니다 (대시보드용).",
        "operationId": "getAllLatestVersions",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "조회할 개수",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LatestVersionsResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/versions/recent-changes": {
      "get": {
        "tags": [
          "Versions"
        ],
        "summary": "최근 변경사항 조회",
        "description": "최근 변경사항을 조회합니다.",
        "operationId": "getRecentChanges",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "description": "조회 기간 (일)",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 7,
              "minimum": 1,
              "maximum": 365
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "최대 개수",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecentChangesResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiVersion": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "MongoDB ObjectId"
          },
          "urlId": {
            "type": "string",
            "description": "API URL의 MongoDB ObjectId"
          },
          "versionId": {
            "type": "string",
            "example": "v2"
          },
          "versionNumber": {
            "type": "integer",
            "example": 2
          },
          "majorVersion": {
            "type": "string",
            "example": "v1"
          },
          "revisionCount": {
            "type": "integer",
            "example": 1
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "lastUpdatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "swaggerJson": {
            "type": "object",
            "description": "Swagger/OpenAPI JSON 객체"
          },
          "changes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Change"
            }
          },
          "changeHistory": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "changesCount": {
                  "type": "integer"
                },
                "summary": {
                  "type": "string"
                }
              }
            }
          },
          "endpointCount": {
            "type": "integer",
            "example": 10
          },
          "parameterCount": {
            "type": "integer",
            "example": 25
          },
          "summary": {
            "type": "string",
            "example": "5개 추가, 2개 삭제"
          }
        }
      },
      "Change": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "added",
              "removed",
              "modified",
              "path_version_changed"
            ]
          },
          "category": {
            "type": "string",
            "enum": [
              "endpoint",
              "parameter",
              "requestBody",
              "response",
              "schema",
              "info",
              "description",
              "server",
              "security",
              "tag",
              "externalDocs",
              "securityScheme",
              "header",
              "example",
              "link",
              "callback"
            ]
          },
          "path": {
            "type": "string",
            "example": "POST /api/users"
          },
          "field": {
            "type": "string",
            "nullable": true
          },
          "oldValue": {
            "type": "object",
            "nullable": true
          },
          "newValue": {
            "type": "object",
            "nullable": true
          },
          "description": {
            "type": "string",
            "example": "사용자 생성 엔드포인트 추가"
          },
          "severity": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "example": "high"
          },
          "recordedAt": {
            "type": "string",
            "format": "date-time"
          },
          "metadata": {
            "type": "object",
            "nullable": true
          }
        }
      },
      "VersionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/ApiVersion"
          }
        }
      },
...
      "LatestVersionsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string"
                },
                "urlId": {
                  "type": "string"
                },
                "versionId": {
                  "type": "string"
                },
                "versionNumber": {
                  "type": "integer"
                },
                "timestamp": {
                  "type": "string",
                  "format": "date-time"
                },
                "endpointCount": {
                  "type": "integer"
                },
                "parameterCount": {
                  "type": "integer"
                },
                "summary": {
                  "type": "string"
                },
                "apiUrl": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "url": {
                      "type": "string"
                    },
                    "group": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "RecentChangesResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiVersion"
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "since": {
                "type": "string",
                "format": "date-time"
              },
              "count": {
                "type": "integer",
                "example": 25
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_ERROR",
                  "NOT_FOUND",
                  "DUPLICATE_URL",
                  "FETCH_ERROR",
                  "INVALID_SWAGGER",
                  "SERVER_ERROR",
                  "RATE_LIMIT_EXCEEDED"
                ],
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string",
                "example": "입력값 검증 실패"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "서비스명은 필수입니다",
                  "URL은 필수입니다"
                ]
              }
            }
          }
        }
      }
    },
    "responses": {
...
      "InternalServerError": {
        "description": "서버 내부 오류 (예상치 못한 오류 발생)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "serverError": {
                "summary": "서버 오류",
                "value": {
                  "success": false,
                  "error": {
                    "code": "SERVER_ERROR",
                    "message": "Database connection failed"
                  }
                }
              },
              "fetchError": {
                "summary": "Swagger JSON 다운로드 실패",
                "value": {
                  "success": false,
                  "error": {
                    "code": "FETCH_ERROR",
                    "message": "Swagger JSON을 가져올 수 없습니다: Network timeout"
                  }
                }
              },
              "invalidSwagger": {
                "summary": "잘못된 Swagger 형식",
                "value": {
                  "success": false,
                  "error": {
                    "code": "INVALID_SWAGGER",
                    "message": "유효하지 않은 Swagger/OpenAPI 형식입니다"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}